Move Gravatar to misc components folder
