+++ src/components/misc/index.js
@@ -1,1 +1,2 @@
 export { default as LinkButton } from 'components/misc/LinkButton'
+export { default as Gravatar } from 'components/misc/Gravatar'
